fix(summary): remove summary page, move logic to correct place
#147
