+++ src/config.js
@@ -2,7 +2,6 @@ import home from './pages/home.vue';
 import create from './pages/create.vue';
 import detail from './pages/detail.vue';
 import teams from './pages/teams.vue';
-import summary from './pages/summary.vue';
 import league from './pages/league.vue';
 import random from './pages/random.vue';
 import knockout from './pages/knockout.vue';
@@ -55,15 +54,6 @@ export const routes = {
     path: 'teams',
     fullPath: '/teams/:slug',
     component: teams,
-    meta: {
-      save: true,
-      fetchCurrentTournament: true
-    }
-  },
-  SUMMARY: {
-    path: 'summary',
-    fullPath: '/summary/:slug',
-    component: summary,
     meta: {
       save: true,
       fetchCurrentTournament: true
