+++ src/pages/random.vue
@@ -129,7 +129,11 @@ export default {
     },
 
     start() {
-      this.$router.push(`/summary/${this.slug}`);
+    if (this.type === 'groupstage') {
+      this.$router.push(`/${this.type}/draw/${this.slug}`);
+    } else {
+      this.$router.push(`/${this.type}/${this.slug}`);
+    }
     }
   },
 };
