+++ src/pages/teams.vue
@@ -29,6 +29,7 @@ export default {
     ...mapGetters({
       slug: 'currentTournament/slug',
       teams: 'currentTournament/teams',
+      type: 'currentTournament/type',
     }),
 
     allTeamsEntered() {
@@ -53,8 +54,16 @@ export default {
     },
 
     submit() {
-      if (this.randomly) this.$router.push(`/random/${this.slug}`);
-      else this.$router.push(`/summary/${this.slug}`);
+      if (this.randomly) {
+        this.$router.push(`/random/${this.slug}`);
+      }
+      else {
+        if (this.type === 'groupstage') {
+          this.$router.push(`/${this.type}/draw/${this.slug}`);
+        } else {
+          this.$router.push(`/${this.type}/${this.slug}`);
+        }
+      }
     }
   }
 };
